saga-kafka-demo/payment-service/requirements.txt
flask==3.0.3
kafka-python==2.3.1